requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
chromedriver-binary-auto==0.3.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
h11==0.14.0
idna==3.10
linkpreview==0.11.0
outcome==1.3.0.post0
psycopg2==2.9.10
PyJWT==2.10.1
pyshorteners==1.0.1
PySocks==1.7.1
requests==2.32.3
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
